Allow pip spec build from build_package.sh
[skip ci]

+++ build_package.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 # Depends on:
 #   UTIL_DIR
-#   REPO_DIR
+#   REPO_DIR | PKG_SPEC
+#       (REPO_DIR for in source build; PKG_SPEC for pip build)
 #   PYTHON_VERSION
 #   BUILD_COMMIT
 #   UNICODE_WIDTHS  (can be empty)
@@ -25,8 +26,16 @@ fi
 # Directory to store wheels
 rm_mkdir /unfixed_wheels
 
-# Enter source tree
-cd /io/$REPO_DIR
+if [ -n "$REPO_DIR" ]; then
+    # Enter source tree
+    cd /io/$REPO_DIR
+    build_source="."
+elif [ -n "$PKG_SPEC" ]; then
+    build_source=$PKG_SPEC
+else:
+    echo "Must specify REPO_DIR or PKG_SPEC"
+    exit 1
+fi
 
 WHEELHOUSE=/io/wheelhouse
 
@@ -36,10 +45,12 @@ for UNICODE_WIDTH in ${UNICODE_WIDTHS}; do
     if [ -n "$BUILD_DEPENDS" ]; then
         $PIP install -f $MANYLINUX_URL $BUILD_DEPENDS
     fi
-    git checkout $BUILD_COMMIT
-    git clean -fxd
-    git reset --hard
-    $PIP wheel -f $MANYLINUX_URL -w /unfixed_wheels --no-deps .
+    if [ -n "$REPO_DIR" ]; then
+        git checkout $BUILD_COMMIT
+        git clean -fxd
+        git reset --hard
+    fi
+    $PIP wheel -f $MANYLINUX_URL -w /unfixed_wheels --no-deps $build_source
 done
 
 # Bundle external shared libraries into the wheels
